Update form validation and error messages
The form validation logic was updated to improve usability. Various error messages were added to the registration form, allowing for more specific feedback on password rules. The method for validating whether an error should be displayed was also moved to a separate function for cleaner code. On input fields, the onTouch event was altered to trigger on input blur instead of value change.