Add validation and error messages to login form
This update includes the addition of the ErrorMessageComponent to the login-page component. The LoginForm email input now requires validation. Error messages will be displayed if the email and password inputs are not valid when attempting to submit the login form. Extraneous console.log statements in the register-page component have also been removed.