Update .gitignore file and allow source control for i18n assets
Changes in this commit include updating the .gitignore file. Some log files are now ignored, while the previous restriction on 'out-tsc' has been lifted. Interestingly, an exception has been added in the file to allow tracking of all files in the 'src/assets/i18n' directory, which might be indispensable for internationalization and localization efforts.