Add SidebarService and update related components
Introduced the SidebarService to manage the display of the sidebar across different components. Toggled the sidebar's visibility based on corresponding views: show when on homepage, hide during the registration or login processes, and update some components' structure to fit this service. Multiple HTML templates, notably in the home, competition, visit, and reservation pages, were adjusted accordingly.