Replace removeUserModal with confirmationModal
This commit removes the specific modal for deleting users and replaces it with a general-purpose confirmation modal that can be reused elsewhere in the application. In addition, it adds a new InvitationsPageComponent which displays a list of available association invites to the user. It provides options to accept or decline the invitations using the new confirmationModal.