Remove big and small input fields, add settings page and associations info
Deleted redundant big and small input field components and their related files. Added new models for handling association statistics and response DTOs for creating a weapon. Added a settings page component to display the information and statistics related to an association. Additionally, introduced a modal for creating a weapon, an interface for handling weapon status and a component for handling default text area inputs.