Add calendar month view component
Added a new CalendarMonthComponent which displays each day and their respective events in a month view format. Along with the component, corresponding HTML template and spec files for testing were created. Also, the component is referenced in the CalendarViewComponent, indicating it's switchable in the main view. Adjustments were made to other unrelated display features in the file of upcoming-events.component.html and calender-view.component.html.