Update API URL in environment configuration
Switched the API url to the azure container app URL and commented out the old one. This change helps to connect with new services deployed in azure container rather than the old server. It reflects the move from our old server to the Azure environment for better scalability and management.