Add track configuration and modal functionality
Added the track configuration feature and the corresponding modal necessary for handling creation, modification, and deletion of tracks. This includes the template, styles, component tests, and component logic for the modals. Improvements were also made to the existing weapon modal, primarily relating to the association ID handling. Routing to the track configuration page feature was included in the app routes.