Add EnrollAtReservationModal for handling reservations
A new EnrollAtReservationModal has been implemented to enable the handling of user enrollments for reservations directly from the calendar. Included changes to the modal functionality in the reservation page component, added necessary inputs and actions in the modal for enrollment, and adjusted related field references for improved consistency. The commit is aimed at enhancing user experience with a simplified and more intuitive reservation process.